+++ src/hkimjp/jpy/scoreboard.clj
@@ -5,7 +5,7 @@
    ; [clojure.java.io :as io]
    [environ.core :refer [env]]
    [org.httpkit.client :as hk-client]
-   [hkimjp.jpy.util :refer [score]]
+   [hkimjp.jpy.util :as u]
    [hkimjp.jpy.view :refer [page]]))
 
 ;; Accept:application/edn
@@ -18,15 +18,20 @@
 (def smile (constantly "😀"))
 
 (defn- submits [author]
-  (map #(-> % first smile) (score author)))
+  (map (fn [[e dt]] [:a {:hx-get (str "/answers/answer/" e)
+                         :hx-target (str "#" author)}
+                     (smile dt)])
+       (u/submits author)))
 
 (defn index [{{:keys [identity]} :session}]
   (page
    [:div.m-4
     [:div.text-2xl.font-medium "scoreboard"]
     (for [user (users)]
-      [:div.flex.mx-4
-       (if (= identity user)
-         [:div {:class "w-24"} [:span.text-white.bg-red-500 user]]
-         [:div {:class "w-24"} user])
-       [:div (submits user)]])])) ;
+      [:div
+       [:div.flex.mx-4
+        (if (= identity user)
+          [:div {:class "w-24"} [:span.text-white.bg-red-500 user]]
+          [:div {:class "w-24"} user])
+        [:div (submits user)]]
+       [:div.mx-4 {:id user}]])]))